Bump o-forms to v6, and o-teaser to v3
Right now, anything that uses this ui library is producing lots of build
warnings with is polluting the dev environment. These errors are comming
from old versions of o-forms and o-teaser, which this lib use.

This commit bumps both of those versions. Though both of them are major
versions, there is no other work required as they are both only major
because of a dependency chain problem which does not affect this library